Revert "refactor: split socials"
This reverts commit 8dfbd83a72b70fca49c473c1b242ac9906c1a038.
