+++ src/components/Socials.tsx
@@ -10,7 +10,35 @@ import "./sidebar/sidebar.css";
 import { IconType } from "react-icons/lib";
 import { SocialAttr } from "../types/DataTypes";
 
-const Socials = ({ contacts }: SocialAttr) => {
+const buildSocialLink = (contact: string, title: string, color: string, Icon: IconType) => {
+  return (
+    <a className={`${color} p-2`} href={contact} target="_blank" rel="noopener noreferrer">
+      <span title={title}>{<Icon size={26} />}</span>
+    </a>
+  );
+};
+
+const buildMobileSocialLink = (contact: string, title: string, color: string, Icon: IconType) => {
+  return (
+    <a className={`${color}`} href={contact} target="_blank" rel="noopener noreferrer">
+      <span title={title}>{<Icon size={26} />}</span>
+    </a>
+  );
+};
+
+const Socials = ({ mobile, contacts }: SocialAttr) => {
+  if (mobile) {
+    return (
+      <div className={`mobile-bio-main mobile-social`}>
+        {buildMobileSocialLink(contacts.linkedin, "LinkedIn", "text-primary", FaLinkedin)}
+        {buildMobileSocialLink(contacts.github, "GitHub", "text-light", FaGithubSquare)}
+        {buildMobileSocialLink(contacts.facebook, "Facebook", "text-info", FaFacebook)}
+        {buildMobileSocialLink(contacts.stackOverFlow, "StackOverFlow", "text-warning", FaStackOverflow)}
+        {buildMobileSocialLink(contacts.resume, "Resume", "text-danger", FaUserGraduate)}
+        {buildMobileSocialLink(contacts.crystal, "CrystalKnows", "text-success", GiCrystalBall)}
+      </div>
+    );
+  }
   return (
     <div className={"side-social-links mt-3"}>
       {buildSocialLink(contacts.linkedin, "LinkedIn", "text-primary", FaLinkedin)}
@@ -20,14 +48,6 @@ const Socials = ({ contacts }: SocialAttr) => {
       {buildSocialLink(contacts.resume, "Resume", "text-danger", FaUserGraduate)}
       {buildSocialLink(contacts.crystal, "CrystalKnows", "text-success", GiCrystalBall)}
     </div>
-  );
-};
-
-const buildSocialLink = (contact: string, title: string, color: string, Icon: IconType) => {
-  return (
-    <a className={`${color} p-2`} href={contact} target="_blank" rel="noopener noreferrer">
-      <span title={title}>{<Icon size={26} />}</span>
-    </a>
   );
 };
